Moved padding to transform.rs, and provided RGBA color option for padding color
